move plugin related functions to plugin.go
